+++ src/app/login/page.tsx
@@ -1,106 +1,16 @@
-"use client"
+export const metadata = {
+  title: "Đăng nhập | KPI",
+  description: "Đăng nhập vào hệ thống quản lý KPI.",
+  keywords: ["đăng nhập", "login", "KPI"],
+  openGraph: {
+    title: "Đăng nhập | KPI",
+    description: "Đăng nhập vào hệ thống quản lý KPI.",
+    type: "website"
+  }
+};
 
-import type React from "react"
-import { useState } from "react"
-import { useRouter } from "next/navigation"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { Alert, AlertDescription } from "@/components/ui/alert"
-import { AlertCircle } from "lucide-react"
-import { login, setAuthToken } from "@/services/auth"
-import { toast } from "sonner"
+import LoginForm from "@/components/login/login-form";
 
 export default function LoginPage() {
-  const router = useRouter()
-  const [email, setEmail] = useState("")
-  const [password, setPassword] = useState("")
-  const [error, setError] = useState("")
-  const [isLoading, setIsLoading] = useState(false)
-
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault()
-
-    if (isLoading) return
-
-    setError("")
-    setIsLoading(true)
-
-    try {
-      if (!email || !password) {
-        setError("Vui lòng nhập email và mật khẩu")
-        setIsLoading(false)
-        return
-      }
-      const response = await login(email, password)
-      setAuthToken(response.token)
-      toast.success("Đăng nhập thành công!")
-      router.push("/dashboard")
-    } catch (error: any) {
-      toast.error("Đăng nhập thất bại")
-      setError(error.response?.data?.message || "Đăng nhập thất bại. Vui lòng thử lại.")
-      setIsLoading(false)
-    }
-  }
-
-  return (
-    <div className="flex h-screen w-full items-center justify-center bg-slate-50">
-      <Card className="w-[400px]">
-        <CardHeader className="space-y-1 text-center">
-          <div className="flex justify-center mb-4">
-            <div className="w-12 h-12 bg-orange-600 rounded-md flex items-center justify-center">
-              <span className="text-white font-bold text-xl">T</span>
-            </div>
-          </div>
-          <CardTitle className="text-2xl font-bold">Đăng nhập</CardTitle>
-          <CardDescription>Nhập thông tin đăng nhập để truy cập hệ thống</CardDescription>
-        </CardHeader>
-        <CardContent className="space-y-4">
-          {error && (
-            <Alert variant="destructive">
-              <AlertCircle className="h-4 w-4" />
-              <AlertDescription>{error}</AlertDescription>
-            </Alert>
-          )}
-          <form onSubmit={handleLogin}>
-            <div className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="email">Email</Label>
-                <Input
-                  id="email"
-                  type="email"
-                  placeholder="Nhập email của bạn"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  disabled={isLoading}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label htmlFor="password">Mật khẩu</Label>
-                <Input
-                  id="password"
-                  type="password"
-                  placeholder="Nhập mật khẩu"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  disabled={isLoading}
-                />
-              </div>
-              <Button 
-                type="submit" 
-                className="w-full bg-orange-600 hover:bg-orange-700"
-                disabled={isLoading}
-              >
-                {isLoading ? "Đang đăng nhập..." : "Đăng nhập"}
-              </Button>
-            </div>
-          </form>
-        </CardContent>
-        <CardFooter className="text-center text-sm text-muted-foreground">
-          <p className="w-full text-center">© {new Date().getFullYear()} Công ty cổ phần Thăng Long</p>
-        </CardFooter>
-      </Card>
-    </div>
-  )
+  return <LoginForm />;
 }
